removed compiler warnings on machines that didn't understand that pid_t is an integer
git-svn-id: http://psotnic.svn.sourceforge.net/svnroot/psotnic@54 642afdc4-8138-0410-ae4e-e573936c0dc0

+++ psotnic-current/class-update.cpp
@@ -113,7 +113,7 @@ bool update::doUpdate(const char *site)
 		if(link.ok() && match("*.tar.gz", link.file) && strlen(link.file) < MAX_LEN)
 		{
 			printf("[+] The newest version is: %s, downloading (this may take a while)\n", link.file);
-			snprintf(updir, 256, ".update-%d", getpid());
+			snprintf(updir, 256, ".update-%d", (int) getpid());
 			
 			if(mkdir(updir, 0700))
 			{
